moving store folder to src
