+++ src/components/Cart.tsx
@@ -1,5 +1,5 @@
 import CartItem from './CartItem'
-import { useAppSelector } from './store/Hooks'
+import { useAppSelector } from '../store/Hooks'
 
 function Cart() {
   const cart = useAppSelector((state) => state.cart)
